Integrate previous metrics comparison
